requirements.txt
langchain==0.1.17
langchain-core==0.1.49
langchain-openai==0.1.6
langchain-community==0.0.36
langserve==0.0.38
pydantic==2.7.4
fastapi==0.111.0
python-dotenv
streamlit
uvicorn
sse-starlette
httpx==0.25.0
